requirements.txt
celery==5.0.5
coverage==5.5
django-admin-interface==0.16.3
django-allauth==0.44.0
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-import-export==2.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.1.2
redis==3.5.3
sentry-sdk==1.0.0
sqlparse==0.4.1
stripe==2.56.0
Unipath==1.1
virtualenv==20.4.4
whitenoise==3.2
honcho==1.0.1
